scripts/requirements.txt
# Bronze Tier Dependencies
# Install with: pip install -r requirements.txt

# File system watching
watchdog>=4.0.0

# Google API dependencies for Gmail integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.97.0

# Scheduling library
schedule>=1.2.0

# WhatsApp Web watcher automation
playwright>=1.48.0
